search: move Zoekt Query conversion test
