+++ internal/search/zoekt/indexed_search_test.go
@@ -5,7 +5,6 @@ import (
 	"crypto/md5"
 	"encoding/binary"
 	"fmt"
-	"sort"
 	"testing"
 	"time"
 
@@ -13,7 +12,6 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/google/zoekt"
-	zoektquery "github.com/google/zoekt/query"
 
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/search"
@@ -463,197 +461,6 @@ func TestZoektResultCountFactor(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestQueryToZoektQuery(t *testing.T) {
-	cases := []struct {
-		Name    string
-		Type    IndexedRequestType
-		Pattern *search.TextPatternInfo
-		Query   string
-	}{
-		{
-			Name: "substr",
-			Type: TextRequest,
-			Pattern: &search.TextPatternInfo{
-				IsRegExp:                     true,
-				IsCaseSensitive:              false,
-				Pattern:                      "foo",
-				IncludePatterns:              nil,
-				ExcludePattern:               "",
-				PathPatternsAreCaseSensitive: false,
-			},
-			Query: "foo case:no",
-		},
-		{
-			Name: "symbol substr",
-			Type: SymbolRequest,
-			Pattern: &search.TextPatternInfo{
-				IsRegExp:                     true,
-				IsCaseSensitive:              false,
-				Pattern:                      "foo",
-				IncludePatterns:              nil,
-				ExcludePattern:               "",
-				PathPatternsAreCaseSensitive: false,
-			},
-			Query: "sym:foo case:no",
-		},
-		{
-			Name: "regex",
-			Type: TextRequest,
-			Pattern: &search.TextPatternInfo{
-				IsRegExp:                     true,
-				IsCaseSensitive:              false,
-				Pattern:                      "(foo).*?(bar)",
-				IncludePatterns:              nil,
-				ExcludePattern:               "",
-				PathPatternsAreCaseSensitive: false,
-			},
-			Query: "(foo).*?(bar) case:no",
-		},
-		{
-			Name: "path",
-			Type: TextRequest,
-			Pattern: &search.TextPatternInfo{
-				IsRegExp:                     true,
-				IsCaseSensitive:              false,
-				Pattern:                      "foo",
-				IncludePatterns:              []string{`\.go$`, `\.yaml$`},
-				ExcludePattern:               `\bvendor\b`,
-				PathPatternsAreCaseSensitive: false,
-			},
-			Query: `foo case:no f:\.go$ f:\.yaml$ -f:\bvendor\b`,
-		},
-		{
-			Name: "case",
-			Type: TextRequest,
-			Pattern: &search.TextPatternInfo{
-				IsRegExp:                     true,
-				IsCaseSensitive:              true,
-				Pattern:                      "foo",
-				IncludePatterns:              []string{`\.go$`, `yaml`},
-				ExcludePattern:               "",
-				PathPatternsAreCaseSensitive: true,
-			},
-			Query: `foo case:yes f:\.go$ f:yaml`,
-		},
-		{
-			Name: "casepath",
-			Type: TextRequest,
-			Pattern: &search.TextPatternInfo{
-				IsRegExp:                     true,
-				IsCaseSensitive:              true,
-				Pattern:                      "foo",
-				IncludePatterns:              []string{`\.go$`, `\.yaml$`},
-				ExcludePattern:               `\bvendor\b`,
-				PathPatternsAreCaseSensitive: true,
-			},
-			Query: `foo case:yes f:\.go$ f:\.yaml$ -f:\bvendor\b`,
-		},
-		{
-			Name: "path matches only",
-			Type: TextRequest,
-			Pattern: &search.TextPatternInfo{
-				IsRegExp:                     true,
-				IsCaseSensitive:              false,
-				Pattern:                      "test",
-				IncludePatterns:              []string{},
-				ExcludePattern:               ``,
-				PathPatternsAreCaseSensitive: true,
-				PatternMatchesContent:        false,
-				PatternMatchesPath:           true,
-			},
-			Query: `f:test`,
-		},
-		{
-			Name: "content matches only",
-			Type: TextRequest,
-			Pattern: &search.TextPatternInfo{
-				IsRegExp:                     true,
-				IsCaseSensitive:              false,
-				Pattern:                      "test",
-				IncludePatterns:              []string{},
-				ExcludePattern:               ``,
-				PathPatternsAreCaseSensitive: true,
-				PatternMatchesContent:        true,
-				PatternMatchesPath:           false,
-			},
-			Query: `c:test`,
-		},
-		{
-			Name: "content and path matches 1",
-			Type: TextRequest,
-			Pattern: &search.TextPatternInfo{
-				IsRegExp:                     true,
-				IsCaseSensitive:              false,
-				Pattern:                      "test",
-				IncludePatterns:              []string{},
-				ExcludePattern:               ``,
-				PathPatternsAreCaseSensitive: true,
-				PatternMatchesContent:        true,
-				PatternMatchesPath:           true,
-			},
-			Query: `test`,
-		},
-		{
-			Name: "content and path matches 2",
-			Type: TextRequest,
-			Pattern: &search.TextPatternInfo{
-				IsRegExp:                     true,
-				IsCaseSensitive:              false,
-				Pattern:                      "test",
-				IncludePatterns:              []string{},
-				ExcludePattern:               ``,
-				PathPatternsAreCaseSensitive: true,
-				PatternMatchesContent:        false,
-				PatternMatchesPath:           false,
-			},
-			Query: `test`,
-		},
-		{
-			Name: "repos must include",
-			Type: TextRequest,
-			Pattern: &search.TextPatternInfo{
-				IsRegExp:                     true,
-				Pattern:                      "foo",
-				FilePatternsReposMustInclude: []string{`\.go$`, `\.yaml$`},
-				FilePatternsReposMustExclude: []string{`\.java$`, `\.xml$`},
-			},
-			Query: `foo (type:repo file:\.go$) (type:repo file:\.yaml$) -(type:repo file:\.java$) -(type:repo file:\.xml$)`,
-		},
-	}
-	for _, tt := range cases {
-		t.Run(tt.Name, func(t *testing.T) {
-			q, err := zoektquery.Parse(tt.Query)
-			if err != nil {
-				t.Fatalf("failed to parse %q: %v", tt.Query, err)
-			}
-			got, err := search.QueryToZoektQuery(tt.Pattern, tt.Type == SymbolRequest)
-			if err != nil {
-				t.Fatal("queryToZoektQuery failed:", err)
-			}
-			if !queryEqual(got, q) {
-				t.Fatalf("mismatched queries\ngot  %s\nwant %s", got.String(), q.String())
-			}
-		})
-	}
-}
-
-func queryEqual(a, b zoektquery.Q) bool {
-	sortChildren := func(q zoektquery.Q) zoektquery.Q {
-		switch s := q.(type) {
-		case *zoektquery.And:
-			sort.Slice(s.Children, func(i, j int) bool {
-				return s.Children[i].String() < s.Children[j].String()
-			})
-		case *zoektquery.Or:
-			sort.Slice(s.Children, func(i, j int) bool {
-				return s.Children[i].String() < s.Children[j].String()
-			})
-		}
-		return q
-	}
-	return zoektquery.Map(a, sortChildren).String() == zoektquery.Map(b, sortChildren).String()
 }
 
 func generateZoektMatches(count int) []zoekt.FileMatch {
